chore: ChannelList/ChannelPreviewAction.tsx
